Moving AggregateDataHandlerBase
