+++ traffic_stops/static/js/app/states/nc/index.js
@@ -1,6 +1,7 @@
 import DataHandlerBase from '../../base/DataHandlerBase.js';
 import VisualBase from '../../base/VisualBase.js';
 import TableBase from '../../base/TableBase.js';
+import AggregateDataHandlerBase from '../../base/AggregateDataHandlerBase.js';
 import Stops from './defaults.js';
 
 import { StopsHandler, StopRatioDonut, StopRatioTimeSeries, StopsTable } from './Stops.js';
@@ -188,28 +189,6 @@ var ContrabandHitRateHandler = DataHandlerBase.extend({
       years: years,
       raw: raw
     });
-  }
-});
-
-var AggregateDataHandlerBase = DataHandlerBase.extend({
-  get_data: function(){
-    var datas = [],
-        checkIfComplete = (data) => {
-          this.numRemaining = this.numRemaining - 1;
-          datas.push(data);
-          if(this.numRemaining === 0){
-            this.set("raw_data", datas);
-            this.set("data", undefined);
-            this.clean_data();
-            this.trigger("dataLoaded", this.get("data"));
-          }
-        };
-
-    this.numRemaining = this.get("handlers").length;
-    _.each(this.get("handlers"), function(handler){
-      this.listenTo(handler, "dataLoaded", checkIfComplete);
-      this.listenTo(handler, "dataRequestFailed", this.showError);
-    }, this);
   }
 });
